requirements.txt
pyinstaller
requests
pyexecjs
schedule
flask